+++ crates/rust-client/src/rpc/tonic_client/mod.rs
@@ -39,7 +39,7 @@ use super::{
     Endpoint, FetchedAccount, NodeRpcClient, RpcEndpoint, NoteSyncInfo, RpcError,
     RpcStatusInfo,
 };
-use crate::rpc::domain::sync::ChainMmrInfo;
+use crate::rpc::domain::sync::{ChainMmrInfo, SyncTarget};
 use crate::rpc::domain::account_vault::{AccountVaultInfo, AccountVaultUpdate};
 use crate::rpc::domain::storage_map::{StorageMapInfo, StorageMapUpdate};
 use crate::rpc::domain::transaction::TransactionsInfo;
@@ -587,18 +587,11 @@ impl NodeRpcClient for GrpcClient {
     async fn sync_chain_mmr(
         &self,
         block_from: BlockNumber,
-        block_to: Option<BlockNumber>,
+        upper_bound: SyncTarget,
     ) -> Result<ChainMmrInfo, RpcError> {
         let block_from = block_from.as_u32();
 
-        let upper_bound = match block_to {
-            Some(block_to) => {
-                Some(proto::rpc::sync_chain_mmr_request::UpperBound::BlockNum(block_to.as_u32()))
-            },
-            None => Some(proto::rpc::sync_chain_mmr_request::UpperBound::ChainTip(
-                proto::rpc::ChainTip::Committed.into(),
-            )),
-        };
+        let upper_bound = Some(upper_bound.into());
 
         let request = proto::rpc::SyncChainMmrRequest { block_from, upper_bound };
 
@@ -885,10 +878,14 @@ impl NodeRpcClient for GrpcClient {
         Ok(proofs)
     }
 
-    async fn get_block_by_number(&self, block_num: BlockNumber) -> Result<ProvenBlock, RpcError> {
+    async fn get_block_by_number(
+        &self,
+        block_num: BlockNumber,
+        include_proof: bool,
+    ) -> Result<ProvenBlock, RpcError> {
         let request = proto::blockchain::BlockRequest {
             block_num: block_num.as_u32(),
-            include_proof: Some(false),
+            include_proof: Some(include_proof),
         };
 
         let response = self